requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.10.0
cachetools==6.2.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
Django==4.2.7
djangorestframework==3.14.0
fastapi==0.135.1
flask==3.0.3
fsspec==2026.2.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
packaging==26.0
postgrest==2.28.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2025.2
realtime==2.28.0
requests==2.32.5
rich==14.3.3
six==1.17.0
sqlparse==0.5.3
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.41.0
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0

